+++ .rubocop.yml
@@ -244,7 +244,7 @@ Lint/AmbiguousAssignment: # (new in 1.7)
 Lint/DeprecatedConstants: # (new in 1.8)
   Enabled: true
 Lint/DuplicateBranch: # (new in 1.3)
-  Enabled: true
+  Enabled: false
 Lint/DuplicateRegexpCharacterClassElement: # (new in 1.1)
   Enabled: true
 Lint/EmptyBlock: # (new in 1.1)
