+++ app/models/forms/dob.rb
@@ -15,7 +15,6 @@ module Forms
     attribute :partner_year, Integer
     attribute :over_16, Integer
 
-
     MINIMUM_AGE = 15
     MAXIMUM_AGE = 120
 
@@ -55,7 +54,8 @@ module Forms
     end
 
     def too_young?
-      return if over_16 == 'false'
+      return false if over_16 == 'false'
+
       date_of_birth > minimum_date_of_birth
     end
