+++ features/support/page_objects/legal_representative_detail_page.rb
@@ -17,5 +17,4 @@ class LegalRepresentativeDetailPage < BasePage
     continue
   end
 
-
 end
